sufuk-getchild&display

+++ 016-GetChild_Display/src/STEPProcessor.h
@@ -63,7 +63,7 @@ public:
 
     void countChildren(const std::shared_ptr<AssemblyNode> &parent);
 
-    unsigned int ShapeCounter;
+    float ShapeCounter;
     unsigned int ProgressOfGetChild;
 
 private:
